created a Utility folder for classes and scripts that are utility. Moved logging functions and classes into their own files based on scope of function. Working on the log function decorator need to create a version that works for functions that do not return a value. Also need to run it in several contexts to create logs and update the decorators according to what would improve the quality of log information
