+++ src/dynode/Utility/custom_log_formatter.py
@@ -1,0 +1,15 @@
+import logging
+
+
+class CustomLogFormatter(logging.Formatter):
+    """ Custom Formatter does these 2 things:
+    1. Overrides 'funcName' with the value of 'func_name_override', if it exists.
+    2. Overrides 'filename' with the value of 'file_name_override', if it exists.
+    """
+
+    def format(self, record):
+        if hasattr(record, 'func_name_override'):
+            record.funcName = record.func_name_override
+        if hasattr(record, 'file_name_override'):
+            record.filename = record.file_name_override
+        return super(CustomLogFormatter, self).format(record)
